Split the updating and checking part into to separate plugins
Setting UNEXPECTED_SNAPSHOT=true will switch to the update plugin
