+++ lib/ensureAfterHookIsRegistered.js
@@ -346,12 +346,10 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
           return replayAndRestoreProcessExit();
         }
 
-        const doUpdate =
-          /^(?:1|true|on|yes)$/i.test(process.env.UNEXPECTED_SNAPSHOT) &&
-          fileNames.length > 0;
         const Promise = require('bluebird');
         const fs = Promise.promisifyAll(require('fs'));
-        if (doUpdate) {
+        console.log(fileNames.length);
+        if (fileNames.length > 0) {
           const results = await Promise.map(
             fileNames,
             fileName =>
@@ -363,13 +361,6 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
             `unexpected-snapshot: Wrote ${results.length} file${
               results.length === 1 ? '' : 's'
             }`
-          );
-        } else {
-          console.log(
-            'UNEXPECTED_SNAPSHOT environment not given, not updating snapshots in source files'
-          );
-          console.log(
-            `Rerun the command with UNEXPECTED_SNAPSHOT=yes to update the snapshots`
           );
         }
         replayAndRestoreProcessExit();
