+++ test/unexpected-snapshot.js
@@ -92,6 +92,7 @@ describe('with snapshot updating on', function() {
         value = beautifyJavaScript(value);
       }
       const tmpFileName = await writeTestToTemporaryFile(subject);
+      expect.errorMode = 'nested';
 
       try {
         const [err, stdout] = await runWithMocha(tmpFileName, {
@@ -612,7 +613,10 @@ describe('with snapshot updating off', function() {
           expect(42, 'to equal snapshot');
         },
         'to throw',
-        'expected 42 to equal snapshot'
+        'expected 42 to equal snapshot\n' +
+          '\n' +
+          'UNEXPECTED_SNAPSHOT environment not given, not updating snapshots in source files\n' +
+          'Rerun the command with UNEXPECTED_SNAPSHOT=yes to update the snapshots'
       );
     });
 
@@ -628,7 +632,10 @@ describe('with snapshot updating off', function() {
           expect(42, 'to equal snapshot', 666);
         },
         'to throw',
-        'expected 42 to equal snapshot 666'
+        'expected 42 to equal snapshot 666\n' +
+          '\n' +
+          'UNEXPECTED_SNAPSHOT environment not given, not updating snapshots in source files\n' +
+          'Rerun the command with UNEXPECTED_SNAPSHOT=yes to update the snapshots'
       );
     });
   });
@@ -640,7 +647,10 @@ describe('with snapshot updating off', function() {
           expect(42, 'to inspect as snapshot');
         },
         'to throw',
-        'expected 42 to inspect as snapshot'
+        'expected 42 to inspect as snapshot\n' +
+          '\n' +
+          'UNEXPECTED_SNAPSHOT environment not given, not updating snapshots in source files\n' +
+          'Rerun the command with UNEXPECTED_SNAPSHOT=yes to update the snapshots'
       );
     });
 
@@ -656,7 +666,13 @@ describe('with snapshot updating off', function() {
           expect(42, 'to inspect as snapshot', '666');
         },
         'to throw',
-        "expected 42 to inspect as snapshot '666'\n" + '\n' + '-42\n' + '+666'
+        "expected 42 to inspect as snapshot '666'\n" +
+          '\n' +
+          '-42\n' +
+          '+666\n' +
+          '\n' +
+          'UNEXPECTED_SNAPSHOT environment not given, not updating snapshots in source files\n' +
+          'Rerun the command with UNEXPECTED_SNAPSHOT=yes to update the snapshots'
       );
     });
   });
